feat(gui): remove `rainbow` colour accent
This is because it is incompatible with the `light-modern` theme. This is probably a warning sign that it would become a maintenance burden as we make more changes to the GUI, so we removed it.

Note that we may reintroduce the accent as a theme, but I can't promise that.